spring boot community day12   增加了回复的底层内容

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -32,7 +32,7 @@ public class PublishController {
             @RequestParam(value = "title" , required = false) String title,
             @RequestParam(value = "description" , required = false) String description,
             @RequestParam(value = "tag" , required = false) String tag,
-            @RequestParam(value = "id" , required = false) Integer id,
+            @RequestParam(value = "id" , required = false) Long id,
             HttpServletRequest request,
             Model model){
         //保存记录
@@ -73,7 +73,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id")Long id,
                        Model model){
         QuestionDTO questionById = questionService.getQuestionById(id);
         //保存记录
